+++ SIRPy.cpp
@@ -13,10 +13,10 @@
 // We also rename enums, so that (e.g.) Python State.SUSCEPTIBLE correponds to C++ SIR::STATE_SUS. If we kept
 // C++ names as-is, Python code would include painful redundandancies (e.g., Event.EVENT_ENTER)
 BOOST_PYTHON_MODULE(PyBoostDemo) {
-	boost::python::class_<SIR>("SIR", boost::python::init<int,int>())
+	boost::python::class_<SIR>("SIR", boost::python::init<size_t,size_t>())
 		.def("initialize", &SIR::initialize)
 		.def<void (SIR::*)(void)>("calculate_proj", &SIR::calculate)
-		.def<void (SIR::*)(int)>("calculate_unit", &SIR::calculate)
+		.def<void (SIR::*)(size_t)>("calculate_unit", &SIR::calculate)
 		.def("state_count", &SIR::state_count)
 		.def("event_count", &SIR::event_count)
 	;
